refactor(api): move learning-content-conversion-service to prescription/shared
