+++ api/src/prescription/shared/domain/services/learning-content-conversion-service.js
@@ -1,4 +1,4 @@
-import * as skillRepository from '../../../../src/shared/infrastructure/repositories/skill-repository.js';
+import * as skillRepository from '../../../../shared/infrastructure/repositories/skill-repository.js';
 
 async function findActiveSkillsForCappedTubes(cappedTubes, dependencies = { skillRepository }) {
   const skills = [];
